requirements.txt
Flask==2.3.2
langchain==0.1.13
langchain_anthropic==0.1.4
langchain_community==0.0.29
langchain_core==0.1.33
langchain_google_genai==0.0.11
langchain_openai==0.1.1
python-dotenv==1.0.1
